requirements.txt
anyio==4.12.0
asgiref==3.11.0
certifi==2025.11.12
Django==5.2.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pillow==12.0.0
python-dotenv==1.2.1
python-telegram-bot==22.5
sqlparse==0.5.4
